feat: getUser 함수 재구현, 사이드 네비게이션 바 데이터 재연동
